### 2021-11-30 #3963 @whedon
[PRE REVIEW]: Kamodo: A functional api for space weather models and data
**Submitting author:** @asherp (<a href="http://orcid.org/0000-0002-5718-1303">Asher Pembroke</a>)
**Repository:** <a href="https://github.com/EnsembleGovServices/kamodo-core" target ="_blank">https://github.com/EnsembleGovServices/kamodo-core</a>
**Version:** 21.10.0
