### 2022-01-10 #3963 @mbobra
Ah I'm so sorry for the late response, @dfm! I'm happy to take over editing duties from you if you wish (you can re-assign me as editor). I am familiar with heliophysics and space weather forecasting so it might be a better fit with @asherp 's work.
