### 2022-01-12 #4053 @whedon
[REVIEW]: Kamodo: A functional api for space weather models and data
**Submitting author:** @asherp (<a href="http://orcid.org/0000-0002-5718-1303">Asher Pembroke</a>)
**Repository:** https://github.com/EnsembleGovServices/kamodo-core
**Branch with paper.md** (empty if default branch):
**Version:** 22.6.0

- [x] **Functionality:** Have the functional claims of the software been confirmed? https://github.com/EnsembleGovServices/kamodo-core/issues/24, https://github.com/EnsembleGovServices/kamodo-core/issues/54

- [x] **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution. https://github.com/EnsembleGovServices/kamodo-core/issues/19
- [x] **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)? https://github.com/EnsembleGovServices/kamodo-core/issues/28, https://github.com/EnsembleGovServices/kamodo-core/issues/58, https://github.com/EnsembleGovServices/kamodo-core/issues/59
- [x] **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified? https://github.com/EnsembleGovServices/kamodo-core/issues/21, https://github.com/EnsembleGovServices/kamodo-core/issues/61
- [x] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support https://github.com/EnsembleGovServices/kamodo-core/issues/20

- [x] **State of the field:** Do the authors describe how this software compares to other commonly-used packages? https://github.com/EnsembleGovServices/kamodo-core/issues/62
- [x] **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax]( https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)? https://github.com/EnsembleGovServices/kamodo-core/issues/62

See below in: https://github.com/openjournals/joss-reviews/issues/4053#issuecomment-1054449513
