### 2022-01-13 EnsembleGovServices/kamodo-core#19 @dstansby
Update dependency list in the docs
[JOSS requirement]
The list of requirements in the docs (https://ccmc.gsfc.nasa.gov/Kamodo/#requirements) is not the same as the requirements in `setup.cfg` (https://github.com/EnsembleGovServices/kamodo-core/blob/2f80f79da1ee9b6c703f2c16eee66e7100311535/setup.cfg#L22). The list in the docs should be updated to reflect the requirements in `setup.cfg`.
