### 2022-01-13 EnsembleGovServices/kamodo-core#20 @dstansby
Add contribution/reporting/support guidelines
[JOSS requirement]
One of the requirements for JOSS is:
> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

As far as I can tell there are no clear guidelines, for this, or they are hard to discover. These guidelines should be added in a prominent position in the docs.
