### 2022-01-13 EnsembleGovServices/kamodo-core#21 @dstansby
Describe how to run (or automate) the test suite
[JOSS requirement]
It looks like there are a series of tests in the repository which is 👍 . These don't seem to be automatically run, and there aren't any instructions in the docs on how to run them. Either:
- Automated runs of the tests should be set up using continuous integration
- Clear steps on how a user can run the test suite should be added to the documentation
