### 2022-01-13 EnsembleGovServices/kamodo-core#24 @dstansby
Fix error message for incompatible units
[JOSS requirement]
The following code errors for me:
```python
from kamodo import Kamodo

kamodo = Kamodo('mass[kg] = x', 'vol[m^3] = y')
kamodo['rho(x,y)[g/cm]'] = 'mass/vol'
```
with:
```python
Traceback (most recent call last):
  File "test.py", line 4, in <module>
    kamodo['rho(x,y)[g/cm]'] = 'mass/vol'
  File "/Users/dstansby/miniconda3/envs/kamodo/lib/python3.7/site-packages/kamodo/kamodo.py", line 653, in __setitem__
    verbose=self.verbose)
  File "/Users/dstansby/miniconda3/envs/kamodo/lib/python3.7/site-packages/kamodo/util.py", line 759, in unify
    verbose=verbose))
  File "/Users/dstansby/miniconda3/envs/kamodo/lib/python3.7/site-packages/kamodo/util.py", line 837, in unify
    raise NameError(error_msg)
UnboundLocalError: local variable 'error_msg' referenced before assignment
```
I deliberately set the units to be incorrect in the second line to check the error message. I don't think this is the intended error message though!
