### 2022-01-25 EnsembleGovServices/kamodo-core#28 @dstansby
Add API documentation
[JOSS requirement]
The online documentation has a really good guide to get users up and running with kamodo. As far as I can tell it doesn't have any API documentation however, ie. a list of all the functions/classes that are provided, along with descriptions of what they do and their inputs/outputs.
