### 2022-01-25 #4053 @dstansby
👋 all. I'm afraid I am away for the next two weeks, so won't get a complete review finished before then. I have done a preliminary pass of the documentation and code, checked off some items, and opened issues (linked in the original post) for other items that need resolving. When I get back I will finish running the examples given in the documentation to check they work as expected.

@samaloney, I would suggest you start by reviewing the paper and providing any feedback on that - that way hopefully a 'complete' set of feedback is available to @asherp to be working on. Sorry again for the delay in my full review.
