### 2022-02-28 #4053 @dfm
@samaloney — sorry about that! We're in the process of transitioning to a new workflow over here. Instead of using the checklist above (I'll actually remove it), can you write a comment that starts with `@editorialbot generate my checklist`? Then the bot will overwrite that comment with a checklist that you _can_ edit, and then we'll use that one. Thanks!
