### 2022-03-01 EnsembleGovServices/kamodo-core#28 @asher-pembroke
This issue has been resolved in and is in the master branch.
