### 2022-03-01 EnsembleGovServices/kamodo-core#24 @asher-pembroke
This has been resolved and pushed to master

```python
from kamodo import Kamodo

kamodo = Kamodo('mass[kg] = x', 'vol[m^3] = y')
kamodo['rho(x,y)[g/cm]'] = 'mass/vol'
```

Now raises the following `NameError`:
```sh
cannot convert mass(x)/vol(y) [kilogram/meter**3] mass/length**3 to rho(x, y)[gram/centimeter] mass/length
```
