### 2022-03-01 EnsembleGovServices/kamodo-core#21 @asher-pembroke
This is now in the master branch README here https://github.com/EnsembleGovServices/kamodo-core#test-suite
