### 2022-03-01 EnsembleGovServices/kamodo-core#20 @asher-pembroke
We've added a CONTRIBUTING.md file to our docs/ available in the documentation site https://ensemblegovservices.github.io/kamodo-core/CONTRIBUTING/
