### 2022-03-01 EnsembleGovServices/kamodo-core#19 @asher-pembroke
Our github pages versions linked from the repo now matches setup.cfg https://ensemblegovservices.github.io/kamodo-core/#requirements We'll ask the ccmc to redirect the outdated documentation site here.
