### 2022-03-02 EnsembleGovServices/kamodo-core#58 @dstansby
Unlcear what some example code is doing
[JOSS requirement]

In the documentation there's the following line:
```python
def rho(x):
    pass
Kamodo(rho = rho)
```
which produces "𝜌(𝑥)=𝜆(𝑥)". It's not clear to me where the lambda is coming from here... Either
- This is a bug that should be fixed
- Where the lambda is coming from should be briefly explained in the docs
