### 2022-03-02 EnsembleGovServices/kamodo-core#28 @dstansby
Could you re-open this? The new API documentation seems to be here: https://ensemblegovservices.github.io/kamodo-core/API/#plotting (let me know if I'm in the wrong place!), but it seems incomplete. In particular:

- There aren't descriptions and types documented for all the input arguments for all methods/functions
- There aren't descriptions and types documented for all the returned values for all methods/functions
