### 2022-03-02 EnsembleGovServices/kamodo-core#28 @dstansby
To automatically generate API documentation from docstrings, I can recommend https://sphinx-automodapi.readthedocs.io/en/latest/index.html
