### 2022-03-02 EnsembleGovServices/kamodo-core#59 @dstansby
sympy link broken in docs
[JOSS requirement]

The sympy link on [this page](https://ensemblegovservices.github.io/kamodo-core/notebooks/Syntax/) is broken - it currently points to https://docs.sympy.org/0.7.0/modules/galgebra/latex_ex/latex_ex.html# which gives me a page not found.
