### 2022-03-02 EnsembleGovServices/kamodo-core#61 @dstansby
Errors when running tests
[JOSS requirement]

I've tried running the test suite, and I'm getting some errors, which should be fixed:

```python
$ pytest --cov kamodo.kamodo --cov kamodo.util --cov plotting kamodo/test_plotting.py kamodo/test_kamodo.py kamodo/test_utils.py
============================================================================== test session starts ===============================================================================
platform darwin -- Python 3.7.12, pytest-7.0.1, pluggy-1.0.0
rootdir: /Users/dstansby/github/kamodo-core
plugins: cov-3.0.0
collected 143 items

kamodo/test_plotting.py .....FF...................                                                                                                                         [ 18%]
kamodo/test_kamodo.py ..............................................................................                                                                       [ 72%]
kamodo/test_utils.py ...................F...................                                                                                                               [100%]

==================================================================================== FAILURES ====================================================================================
___________________________________________________________________________ test_vector_plot_3d_vector ___________________________________________________________________________

    def test_vector_plot_3d_vector():
        """
        Method for testing 3D vector plot's trace, layout and plot type.
        """
        x, y, z = np.meshgrid(np.linspace(-2, 2, 4),
                              np.linspace(-3, 3, 6),
                              np.linspace(-5, 5, 10))
        points = np.array(list(zip(x.ravel(), y.ravel(), z.ravel())))

        def fvec_Ncomma3(rvec_Ncomma3=points):
            return rvec_Ncomma3

        title = '$\\vec{f}_{N,3}{\\left(\\vec{r}_{N,3} \\right)} = \\lambda{\\left(\\vec{r}_{N,3} \\right)}$'
        kamodo = Kamodo(fvec_Ncomma3=fvec_Ncomma3)

        results = kamodo.evaluate('fvec_Ncomma3', rvec_Ncomma3=points)

        titles = dict(variable='fvec_Ncomma3', title=title)
>       [trace], plot_type, layout = vector_plot(results, titles, True)

kamodo/test_plotting.py:144:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
kamodo/plotting.py:325: in vector_plot
    cmin=0, cmax=norms.max(),
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

a = array([6.164414  , 5.30315536, 4.5514887 , 3.9721251 , 3.64810115,
       3.64810115, 3.9721251 , 4.5514887 , 5.303155...  , 5.30315536, 4.5514887 , 3.9721251 , 3.64810115,
       3.64810115, 3.9721251 , 4.5514887 , 5.30315536, 6.164414  ])
axis = None, out = None, keepdims = False, initial = <no value>, where = True

    def _amax(a, axis=None, out=None, keepdims=False,
              initial=_NoValue, where=True):
>       return umr_maximum(a, axis, None, out, keepdims, initial, where)
E       TypeError: float() argument must be a string or a number, not '_NoValueType'

../../mambaforge/envs/kamodo/lib/python3.7/site-packages/numpy/core/_methods.py:40: TypeError
------------------------------------------------------------------------------ Captured stdout call ------------------------------------------------------------------------------
	 3-d output (240, 3) (240, 3)
	 3d vector plot
_________________________________________________________________________ test_vector_plot_3d_vector_pd __________________________________________________________________________

    def test_vector_plot_3d_vector_pd():
        x, y, z = np.meshgrid(np.linspace(-2, 2, 4),
                              np.linspace(-3, 3, 6),
                              np.linspace(-5, 5, 10))
        points = pd.DataFrame(list(zip(x.ravel(), y.ravel(), z.ravel())))

        def fvec_Ncomma3(rvec_Ncomma3=points):
            return rvec_Ncomma3

        title = '$\\vec{f}_{N,3}{\\left(\\vec{r}_{N,3} \\right)} = \\lambda{\\left(\\vec{r}_{N,3} \\right)}$'
        kamodo = Kamodo(fvec_Ncomma3=fvec_Ncomma3)

        results = kamodo.evaluate('fvec_Ncomma3', rvec_Ncomma3=points)

        titles = dict(variable='fvec_Ncomma3', title=title)

>       [trace], plot_type, layout = vector_plot(results, titles, True)

kamodo/test_plotting.py:167:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
kamodo/plotting.py:325: in vector_plot
    cmin=0, cmax=norms.max(),
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

a = array([6.164414  , 5.30315536, 4.5514887 , 3.9721251 , 3.64810115,
       3.64810115, 3.9721251 , 4.5514887 , 5.303155...  , 5.30315536, 4.5514887 , 3.9721251 , 3.64810115,
       3.64810115, 3.9721251 , 4.5514887 , 5.30315536, 6.164414  ])
axis = None, out = None, keepdims = False, initial = <no value>, where = True

    def _amax(a, axis=None, out=None, keepdims=False,
              initial=_NoValue, where=True):
>       return umr_maximum(a, axis, None, out, keepdims, initial, where)
E       TypeError: float() argument must be a string or a number, not '_NoValueType'

../../mambaforge/envs/kamodo/lib/python3.7/site-packages/numpy/core/_methods.py:40: TypeError
------------------------------------------------------------------------------ Captured stdout call ------------------------------------------------------------------------------
	 3-d output (240, 3) (240, 3)
	 3d vector plot
__________________________________________________________________________________ test_pad_nan __________________________________________________________________________________

array = array([[1, 2],
       [1, 2]])

    def pad_nan(array):
        if isinstance(array, types.GeneratorType):
            return np.array(list(array))
        try:
            if len(array.shape) == 1:
                # print('padding array {}'.format(array.shape))
                return np.pad(array.astype(float), (0, 1),
                              mode='constant', constant_values=np.nan).T
            elif len(array.shape) == 2:
                # print('padding array {}'.format(array.shape))
                return np.pad(array.astype(float), ((0, 1), (0, 0)),
>                             mode='constant', constant_values=np.nan)

kamodo/util.py:334:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

args = (array([[1., 2.],
       [1., 2.]]), ((0, 1), (0, 0))), kwargs = {'constant_values': nan, 'mode': 'constant'}, relevant_args = (array([[1., 2.],
       [1., 2.]]),)

>   ???

<__array_function__ internals>:6:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

array = array([[1., 2.],
       [1., 2.]]), pad_width = array([[0, 1],
       [0, 0]]), mode = 'constant', kwargs = {'constant_values': nan}

    @array_function_dispatch(_pad_dispatcher, module='numpy')
    def pad(array, pad_width, mode='constant', **kwargs):
        """
        Pad an array.

        Parameters
        ----------
        array : array_like of rank N
            The array to pad.
        pad_width : {sequence, array_like, int}
            Number of values padded to the edges of each axis.
            ((before_1, after_1), ... (before_N, after_N)) unique pad widths
            for each axis.
            ((before, after),) yields same before and after pad for each axis.
            (pad,) or int is a shortcut for before = after = pad width for all
            axes.
        mode : str or function, optional
            One of the following string values or a user supplied function.

            'constant' (default)
                Pads with a constant value.
            'edge'
                Pads with the edge values of array.
            'linear_ramp'
                Pads with the linear ramp between end_value and the
                array edge value.
            'maximum'
                Pads with the maximum value of all or part of the
                vector along each axis.
            'mean'
                Pads with the mean value of all or part of the
                vector along each axis.
            'median'
                Pads with the median value of all or part of the
                vector along each axis.
            'minimum'
                Pads with the minimum value of all or part of the
                vector along each axis.
            'reflect'
                Pads with the reflection of the vector mirrored on
                the first and last values of the vector along each
                axis.
            'symmetric'
                Pads with the reflection of the vector mirrored
                along the edge of the array.
            'wrap'
                Pads with the wrap of the vector along the axis.
                The first values are used to pad the end and the
                end values are used to pad the beginning.
            'empty'
                Pads with undefined values.

                .. versionadded:: 1.17

            <function>
                Padding function, see Notes.
        stat_length : sequence or int, optional
            Used in 'maximum', 'mean', 'median', and 'minimum'.  Number of
            values at edge of each axis used to calculate the statistic value.

            ((before_1, after_1), ... (before_N, after_N)) unique statistic
            lengths for each axis.

            ((before, after),) yields same before and after statistic lengths
            for each axis.

            (stat_length,) or int is a shortcut for before = after = statistic
            length for all axes.

            Default is ``None``, to use the entire axis.
        constant_values : sequence or scalar, optional
            Used in 'constant'.  The values to set the padded values for each
            axis.

            ``((before_1, after_1), ... (before_N, after_N))`` unique pad constants
            for each axis.

            ``((before, after),)`` yields same before and after constants for each
            axis.

            ``(constant,)`` or ``constant`` is a shortcut for ``before = after = constant`` for
            all axes.

            Default is 0.
        end_values : sequence or scalar, optional
            Used in 'linear_ramp'.  The values used for the ending value of the
            linear_ramp and that will form the edge of the padded array.

            ``((before_1, after_1), ... (before_N, after_N))`` unique end values
            for each axis.

            ``((before, after),)`` yields same before and after end values for each
            axis.

            ``(constant,)`` or ``constant`` is a shortcut for ``before = after = constant`` for
            all axes.

            Default is 0.
        reflect_type : {'even', 'odd'}, optional
            Used in 'reflect', and 'symmetric'.  The 'even' style is the
            default with an unaltered reflection around the edge value.  For
            the 'odd' style, the extended part of the array is created by
            subtracting the reflected values from two times the edge value.

        Returns
        -------
        pad : ndarray
            Padded array of rank equal to `array` with shape increased
            according to `pad_width`.

        Notes
        -----
        .. versionadded:: 1.7.0

        For an array with rank greater than 1, some of the padding of later
        axes is calculated from padding of previous axes.  This is easiest to
        think about with a rank 2 array where the corners of the padded array
        are calculated by using padded values from the first axis.

        The padding function, if used, should modify a rank 1 array in-place. It
        has the following signature::

            padding_func(vector, iaxis_pad_width, iaxis, kwargs)

        where

            vector : ndarray
                A rank 1 array already padded with zeros.  Padded values are
                vector[:iaxis_pad_width[0]] and vector[-iaxis_pad_width[1]:].
            iaxis_pad_width : tuple
                A 2-tuple of ints, iaxis_pad_width[0] represents the number of
                values padded at the beginning of vector where
                iaxis_pad_width[1] represents the number of values padded at
                the end of vector.
            iaxis : int
                The axis currently being calculated.
            kwargs : dict
                Any keyword arguments the function requires.

        Examples
        --------
        >>> a = [1, 2, 3, 4, 5]
        >>> np.pad(a, (2, 3), 'constant', constant_values=(4, 6))
        array([4, 4, 1, ..., 6, 6, 6])

        >>> np.pad(a, (2, 3), 'edge')
        array([1, 1, 1, ..., 5, 5, 5])

        >>> np.pad(a, (2, 3), 'linear_ramp', end_values=(5, -4))
        array([ 5,  3,  1,  2,  3,  4,  5,  2, -1, -4])

        >>> np.pad(a, (2,), 'maximum')
        array([5, 5, 1, 2, 3, 4, 5, 5, 5])

        >>> np.pad(a, (2,), 'mean')
        array([3, 3, 1, 2, 3, 4, 5, 3, 3])

        >>> np.pad(a, (2,), 'median')
        array([3, 3, 1, 2, 3, 4, 5, 3, 3])

        >>> a = [[1, 2], [3, 4]]
        >>> np.pad(a, ((3, 2), (2, 3)), 'minimum')
        array([[1, 1, 1, 2, 1, 1, 1],
               [1, 1, 1, 2, 1, 1, 1],
               [1, 1, 1, 2, 1, 1, 1],
               [1, 1, 1, 2, 1, 1, 1],
               [3, 3, 3, 4, 3, 3, 3],
               [1, 1, 1, 2, 1, 1, 1],
               [1, 1, 1, 2, 1, 1, 1]])

        >>> a = [1, 2, 3, 4, 5]
        >>> np.pad(a, (2, 3), 'reflect')
        array([3, 2, 1, 2, 3, 4, 5, 4, 3, 2])

        >>> np.pad(a, (2, 3), 'reflect', reflect_type='odd')
        array([-1,  0,  1,  2,  3,  4,  5,  6,  7,  8])

        >>> np.pad(a, (2, 3), 'symmetric')
        array([2, 1, 1, 2, 3, 4, 5, 5, 4, 3])

        >>> np.pad(a, (2, 3), 'symmetric', reflect_type='odd')
        array([0, 1, 1, 2, 3, 4, 5, 5, 6, 7])

        >>> np.pad(a, (2, 3), 'wrap')
        array([4, 5, 1, 2, 3, 4, 5, 1, 2, 3])

        >>> def pad_with(vector, pad_width, iaxis, kwargs):
        ...     pad_value = kwargs.get('padder', 10)
        ...     vector[:pad_width[0]] = pad_value
        ...     vector[-pad_width[1]:] = pad_value
        >>> a = np.arange(6)
        >>> a = a.reshape((2, 3))
        >>> np.pad(a, 2, pad_with)
        array([[10, 10, 10, 10, 10, 10, 10],
               [10, 10, 10, 10, 10, 10, 10],
               [10, 10,  0,  1,  2, 10, 10],
               [10, 10,  3,  4,  5, 10, 10],
               [10, 10, 10, 10, 10, 10, 10],
               [10, 10, 10, 10, 10, 10, 10]])
        >>> np.pad(a, 2, pad_with, padder=100)
        array([[100, 100, 100, 100, 100, 100, 100],
               [100, 100, 100, 100, 100, 100, 100],
               [100, 100,   0,   1,   2, 100, 100],
               [100, 100,   3,   4,   5, 100, 100],
               [100, 100, 100, 100, 100, 100, 100],
               [100, 100, 100, 100, 100, 100, 100]])
        """
        array = np.asarray(array)
        pad_width = np.asarray(pad_width)

        if not pad_width.dtype.kind == 'i':
            raise TypeError('`pad_width` must be of integral type.')

        # Broadcast to shape (array.ndim, 2)
>       pad_width = _as_pairs(pad_width, array.ndim, as_index=True)

../../mambaforge/envs/kamodo/lib/python3.7/site-packages/numpy/lib/arraypad.py:743:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

x = array([[0, 1],
       [0, 0]]), ndim = 2, as_index = True

    def _as_pairs(x, ndim, as_index=False):
        """
        Broadcast `x` to an array with the shape (`ndim`, 2).

        A helper function for `pad` that prepares and validates arguments like
        `pad_width` for iteration in pairs.

        Parameters
        ----------
        x : {None, scalar, array-like}
            The object to broadcast to the shape (`ndim`, 2).
        ndim : int
            Number of pairs the broadcasted `x` will have.
        as_index : bool, optional
            If `x` is not None, try to round each element of `x` to an integer
            (dtype `np.intp`) and ensure every element is positive.

        Returns
        -------
        pairs : nested iterables, shape (`ndim`, 2)
            The broadcasted version of `x`.

        Raises
        ------
        ValueError
            If `as_index` is True and `x` contains negative elements.
            Or if `x` is not broadcastable to the shape (`ndim`, 2).
        """
        if x is None:
            # Pass through None as a special case, otherwise np.round(x) fails
            # with an AttributeError
            return ((None, None),) * ndim

        x = np.array(x)
        if as_index:
            x = np.round(x).astype(np.intp, copy=False)

        if x.ndim < 3:
            # Optimization: Possibly use faster paths for cases where `x` has
            # only 1 or 2 elements. `np.broadcast_to` could handle these as well
            # but is currently slower

            if x.size == 1:
                # x was supplied as a single value
                x = x.ravel()  # Ensure x[0] works for x.ndim == 0, 1, 2
                if as_index and x < 0:
                    raise ValueError("index can't contain negative values")
                return ((x[0], x[0]),) * ndim

            if x.size == 2 and x.shape != (2, 1):
                # x was supplied with a single value for each side
                # but except case when each dimension has a single value
                # which should be broadcasted to a pair,
                # e.g. [[1], [2]] -> [[1, 1], [2, 2]] not [[1, 2], [1, 2]]
                x = x.ravel()  # Ensure x[0], x[1] works
                if as_index and (x[0] < 0 or x[1] < 0):
                    raise ValueError("index can't contain negative values")
                return ((x[0], x[1]),) * ndim

>       if as_index and x.min() < 0:

../../mambaforge/envs/kamodo/lib/python3.7/site-packages/numpy/lib/arraypad.py:513:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

a = array([[0, 1],
       [0, 0]]), axis = None, out = None, keepdims = False, initial = <no value>, where = True

    def _amin(a, axis=None, out=None, keepdims=False,
              initial=_NoValue, where=True):
>       return umr_minimum(a, axis, None, out, keepdims, initial, where)
E       TypeError: int() argument must be a string, a bytes-like object or a number, not '_NoValueType'

../../mambaforge/envs/kamodo/lib/python3.7/site-packages/numpy/core/_methods.py:44: TypeError

    def test_pad_nan():
        array = np.array([[1, 2], [1, 2]])
>       solution = pad_nan(array)

kamodo/test_utils.py:296:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

array = array([[1, 2],
       [1, 2]])

    def pad_nan(array):
        if isinstance(array, types.GeneratorType):
            return np.array(list(array))
        try:
            if len(array.shape) == 1:
                # print('padding array {}'.format(array.shape))
                return np.pad(array.astype(float), (0, 1),
[…]
