### 2022-03-02 EnsembleGovServices/kamodo-core#62 @dstansby
Add refs to previous 
[JOSS requirement]

In the JOSS paper there is the sentence:
> Several attempts have been made to unify all existing space weather resources around common standards, but have met with limited success.

This statement should be backed up with a couple of examples of such previous attempts, ideally with references for readers who want to find out more about previous attempts.
