### 2022-03-07 EnsembleGovServices/kamodo-core#54 @asher-pembroke
The problem was this line

```python
points = np.array(zip(xx.ravel(), yy.ravel()))
```
In the python 2.7 days, this would have created a 2d array. Now you have to cast the zip into an array first

```python
points = np.array(list(zip(xx.ravel(), yy.ravel())))
```
After that, the rest of the code works. I'll update the docs to reflect this.
