### 2022-03-08 EnsembleGovServices/kamodo-core#58 @asher-pembroke
We've updated the documentation to clarify our use of lambda functions https://ensemblegovservices.github.io/kamodo-core/notebooks/Kamodo/#lambda-functions
