### 2022-03-08 EnsembleGovServices/kamodo-core#59 @asher-pembroke
Working on this. The new sympy docs no longer contains a full explanation of how their parser works. The old documentation was much better!

http://horacio9573.no-ip.org/sympy-0.6.6-docs-html/modules/galgebra/latex_ex/latex_ex.html
