### 2022-03-08 EnsembleGovServices/kamodo-core#59 @asher-pembroke
Fixed the link in the documentation site, added it to the joss branch. Closing for now.

https://ensemblegovservices.github.io/kamodo-core/notebooks/Syntax/
