### 2022-03-09 EnsembleGovServices/kamodo-core#61 @asher-pembroke
Did you install from pypi or from master? We were waiting to merge all the joss fixes before the next public release.

from the base of the git repo: `pip install -e .`
