### 2022-03-22 EnsembleGovServices/kamodo-core#28 @asher-pembroke
We've made some significant improvements to our api recently. Closing for now.
