### 2022-04-06 EnsembleGovServices/kamodo-core#62 @asher-pembroke
We have added a section on our previous work with Kameleon which was an attempt to standardize space weather model output.
