### 2022-06-15 #4053 @samaloney
Sorry for the delay looks ok to me too.

Shouldn't hold up the review but trying to run the FieldIntegration.ipynb example gives an error about missing dependancy `kaleido`.  While just commenting out the `pio.write_image(fig, '....')` call allows the notebook to run there are no figure outputs as there are no `fig.show()` calls, same in the other notebooks too.
