### 2022-06-28 #4053 @lrastaet
Please correct the reference for the Weierstrass function:
Correct spelling of "Werth", "Differentialquotienten" in the title and supply the full title of the book and full name and location of the publisher.

Karl Weierstrass, ["Über continuirliche Functionen eines reellen Arguments, die für keinen Werth des letzteren einen bestimmten Differentialquotienten besitzen,"](https://books.google.com/books?id=1FhtAAAAMAAJ&pg=PA71#v=onepage&q&f=false) (On continuous functions of a real argument which possess a definite derivative for no value of the argument) in: Königlich Preussichen Akademie der Wissenschaften, Mathematische Werke von Karl Weierstrass (Berlin, Germany: Mayer & Mueller, 1895), vol. 2, pages 71–74.;
