### 2022-07-08 #4053 @danielskatz
👋 @lrastaet, @asherp / @asher-pembroke - how does the new proof look?  Are we ready to publish?
