### 2022-07-14 #4053 @arfon
> On page 2 of the manuscript ('Usage' chapter) the computer code shown does not yet have the corrected Weierstrass citation ( spelling in the title of the work, full book title and publisher as in the References). @asher-pembroke / @asherp - your updated version you sent me on 7/6 is OK. Please submit that one.

@asher-pembroke @lrastaet – is this fixed now? If so, can we go ahead and publish?
